parent-monitoring-system/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1

motor==3.3.2
pymongo==4.6.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

python-multipart==0.0.9

pydantic==2.6.1
pydantic-settings==2.1.0

email-validator==2.1.0